WIP: neighbors synchronization
